Added confirmation messages

+++ helper/delete.go
@@ -1,6 +1,8 @@
 package helper
 
 import (
+	"bufio"
+	"fmt"
 	"os"
 	"path/filepath"
 )
@@ -23,23 +25,44 @@ func DeleteDirectory(src string) error {
 	return nil
 }
 
-// Delete deletes a file or directory from src.
-func Delete(src string) error {
+// Delete deletes a file or directory from src. If force is true, it will delete without confirmation.
+func Delete(src string, force bool) error {
 	matches, err := filepath.Glob(src)
 	if err != nil {
 		return err
 	}
 
 	run := func(info os.FileInfo, match string) error {
-		if info.IsDir() {
-			err = DeleteDirectory(match)
+		var input string
+
+		if !force {
+			reader := bufio.NewReader(os.Stdin)
+
+			fmt.Printf("Are you sure you want to delete %s? (y/n): ", match)
+			input, err = reader.ReadString('\n')
+			if err != nil {
+				fmt.Println("Error reading input:", err)
+				return err
+			}
+		}
+
+		if input == "y\n" || input == "Y\n" || force {
+			if info.IsDir() {
+				err = DeleteDirectory(match)
+			} else {
+				err = DeleteFile(match)
+			}
 		} else {
-			err = DeleteFile(match)
+			return fmt.Errorf("deletion of %s cancelled", match)
 		}
 		return err
 	}
 
 	err = RunConcurrent(run, 4, matches)
+
+	if err != nil {
+		return err
+	}
 
 	return nil
 }
